deploy v1.1.1

+++ core/handlers/media_group_handlers.py
@@ -1,35 +1,46 @@
 import json
+import logging
 
 from aiogram import Bot, Router, F
 from aiogram.types import Message
 from aiogram.fsm.context import FSMContext
 
-from core.keyboards import choice_keyboard
+from core.keyboards import date_keyboard
 from core.models import PostStates
 
 
 async def forward_media_to_admin_chat(message: Message, bot: Bot, album: list[Message], state: FSMContext) -> None:
     """This handler will forward a complete album of any type."""
+    logging.debug('forward_media_to_admin_chat')
+
     print(message.model_dump_json())
-    # need special method to serialise Aiogram message .model_dump_json(), json.dumps do incorrect result
-    json_obj = json.loads(message.model_dump_json())
 
     data = await state.get_data()
 
-    if not album:
-        json_album = [json.loads(message.model_dump_json())]
-    else:
-        json_album = [json.loads(message.model_dump_json()) for message in album]
+    json_album = [json.loads(message.model_dump_json()) for message in album]
+    exceed_captions = [len(message.caption) - 1024 for message in album if message.caption]
+
+    print(exceed_captions)
+
+    if max(exceed_captions) > 0:
+        await bot.send_message(chat_id=message.chat.id,
+                               text=f'Cократи пост на {max(exceed_captions)} символов',
+                               reply_to_message_id=message.message_id)
+        return
 
     # delete previous bot msg
     await bot.delete_message(chat_id=data['answer_msg_chat_id'], message_id=data['answer_msg_id'])
     # send new one
     msg = await bot.send_message(chat_id=data['answer_msg_chat_id'],
-                                 text='Выбери пост или реклама?',
-                                 reply_markup=choice_keyboard()
+                                 text='Дата размещения?',
+                                 reply_markup=date_keyboard()
                                  )
+    # msg = await bot.send_message(chat_id=data['answer_msg_chat_id'],
+    #                              text='Выбери пост или реклама?',
+    #                              reply_markup=choice_keyboard()
+    #                              )
     # set state to handle callback
-    await state.set_state(PostStates.POST_OR_ADVERTISEMENT)
+    await state.set_state(PostStates.DATE)
     # add state data
     await state.update_data(answer_msg_id=msg.message_id,
                             answer_msg_chat_id=msg.chat.id,
